fix(example): bypass Cloudflare protection error by adding browser-like headers & separate posts screen
